Added offset non-straight curve functionality to OCC according to ACIS implementation. Currently code works this way:
Straight curve will offset according to offset-distance and offset-direction.

A non-straight curve will not honor offset-direction, rather give a warning of it, just as ACIS.

All input curves will be chained to form a wire to offset, if the wire is not obtainable for all curves, code will error out as in ACIS.

Closed curve(s) will be offset based on distance, positive distance will offset to outside and negative distance will offset to inside. The number of curves may be more than the original input if it's a closed spline.

Open curve(s) will be tricky, since the opencascade code offset the curves both ways and connected the end into a closed wire. CGM has to determine which curves are the needed, hopefully it works out correctly. However, if original curve causes more than one offset curves in the same direction, code will be errored out.

Passed new added tests, will be committed.




git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/cgm/trunk@5356 6162379e-bd28-0410-9a7d-b7f4dcfcad3c