Progress on building with CMake.
CGM now builds against some versions of OpenCascade.
Many horrible things are done to make it work and it
is ever-more-unclear to me why the subdirectories are
separate projects instead of a top-level CMakeLists that
prepares "add_definitions" and "include_directories"
values for all subdirectories.

CGM also creates `cgm.make`, but it is not quite usable.